Move postgres type to its own file
